requirements.txt
# ─────────────────────────────────────────
# Web Framework
# ─────────────────────────────────────────
fastapi>=0.115.0
uvicorn>=0.34.0
pydantic-settings>=2.7.0

# ─────────────────────────────────────────
# Database
# ─────────────────────────────────────────
sqlalchemy>=2.0.0
alembic>=1.14.0
asyncpg>=0.29.0          # PostgreSQL async driver
pgvector>=0.3.0          # PostgreSQL vector extension

# ─────────────────────────────────────────
# Cache
# ─────────────────────────────────────────
redis[asyncio]>=5.0.0

# ─────────────────────────────────────────
# Authentication
# ─────────────────────────────────────────
python-jose[cryptography]>=3.3.0
cryptography>=42.0.0

# ─────────────────────────────────────────
# AI / LLM
# ─────────────────────────────────────────
openai>=1.75.0
anthropic>=0.40.0
langchain>=0.3.0
langchain-openai>=0.3.0
langgraph>=0.3.0
langsmith>=0.2.0

# ─────────────────────────────────────────
# HTTP / Scraping
# ─────────────────────────────────────────
httpx>=0.27.0
beautifulsoup4>=4.12.0
youtube-transcript-api>=0.6.0

# ─────────────────────────────────────────
# NLP
# ─────────────────────────────────────────
kiwipiepy>=0.17.0
numpy>=2.0.0

# ─────────────────────────────────────────
# Scheduler
# ─────────────────────────────────────────
apscheduler>=3.10.0

# ─────────────────────────────────────────
# Market Data
# ─────────────────────────────────────────
yfinance>=0.2.0

# ─────────────────────────────────────────
# Calendar / Holidays
# ─────────────────────────────────────────
holidays>=0.50
